fix tests after io changes
