+++ test/queries/informations_tests.jl
@@ -4,12 +4,11 @@ using ProbabilisticCircuits
 
 @testset "Entropy and KLD" begin
 
-    pc1, vtree = load_struct_prob_circuit(
-                    zoo_psdd_file("simple2.1.psdd"), zoo_vtree_file("simple2.vtree"))
-    pc2, vtree = load_struct_prob_circuit(
-                    zoo_psdd_file("simple2.2.psdd"), zoo_vtree_file("simple2.vtree"))
-    pc3, vtree = load_struct_prob_circuit(
-                    zoo_psdd_file("simple2.3.psdd"), zoo_vtree_file("simple2.vtree"))
+    vtree_file = zoo_vtree_file("simple2.vtree")
+
+    pc1 = read((zoo_psdd_file("simple2.1.psdd"), vtree_file), StructProbCircuit)
+    pc2 = read((zoo_psdd_file("simple2.2.psdd"), vtree_file), StructProbCircuit)
+    pc3 = read((zoo_psdd_file("simple2.3.psdd"), vtree_file), StructProbCircuit)
    
     @test entropy(pc1) ≈ 1.2899219826090118
     @test entropy(pc2) ≈ 0.9359472745536583
